requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
faiss-cpu
google-generativeai
pypdf
